requirements.txt
opencv-python
numpy
matplotlib
pillow
ttkbootstrap
